feat: Refactor of DoofinderApi + creation of DoofinderAdminPanelView class
