+++ lib/DoofinderException.php
@@ -4,10 +4,13 @@
  * @copyright Doofinder
  * @license   GPLv3
  */
+
+namespace PrestaShop\Module\Doofinder\Lib;
+
 if (!defined('_PS_VERSION_')) {
     exit;
 }
 
-class DoofinderException extends Exception
+class DoofinderException extends \Exception
 {
 }
